Move GCP tests into dedicated modules
Extracts the test modules from gcp/mod.rs and gcp/federation/mod.rs
into gcp/tests.rs and gcp/federation/tests.rs. Pure move plus the
rustfmt reflow that the one-level dedent triggers; no test logic
changes.
